requirements.txt
atomicwrites==1.3.0
attrs==19.1.0
colorama==0.4.1
dj-database-url==0.5.0
Django==2.2.6
django-annoying==0.10.4
django-heroku==0.3.1
django-rest-framework==0.1.0
djangorestframework==3.10.1
gunicorn==19.9.0
importlib-metadata==0.20
more-itertools==7.2.0
packaging==19.1
pluggy==0.12.0
psycopg2==2.8.3
py==1.8.0
pyparsing==2.4.2
pytest==5.1.2
python-dotenv==0.10.3
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
wcwidth==0.1.7
whitenoise==4.1.4
zipp==0.6.0
